requirements.txt
fastapi
uvicorn
crawl4ai
openai
supabase
python-dotenv
pydantic
pydantic-settings
httpx
